+++ test/hexapdf/test_revisions.rb
@@ -118,6 +118,7 @@ describe HexaPDF::Revisions do
 
   describe "initialize" do
     it "automatically loads all revisions from the underlying IO object" do
+      assert_kind_of(HexaPDF::Parser, @revisions.parser)
       assert_equal(20, @revisions.revision(0).object(2).value)
       assert_equal(300, @revisions[1].object(2).value)
       assert_equal(400, @revisions[2].object(2).value)